requirements.txt
# SambaNova specific
sseclient-py>=1.7.2

# Core dependencies
llama-index-core>=0.10.0
llama-index-llms-sambanovasystems>=0.4.3
llama-index-embeddings-huggingface>=0.1.6
llama-index-vector-stores-qdrant>=0.1.3
llama-index-readers-file>=0.1.3

# Vector store and embeddings
qdrant-client>=1.1.7
sentence-transformers>=2.2.2
pypdf>=3.0.0

# API and Server
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-magic>=0.4.27

# UI dependencies
streamlit>=1.24.0
rich>=13.7.0
typer>=0.9.0
prompt-toolkit>=3.0.43

# ML dependencies
torch>=2.0.1
transformers>=4.30.2
numpy>=1.24.3

# Utilities
loguru>=0.7.0
pathlib
typing-extensions>=4.5.0
requests>=2.31.0
colorama>=0.4.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# OpenAI
openai>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-magic>=0.4.27
openpyxl

# Plotly
plotly>=5.18.0
networkx>=3.2.1
scikit-learn>=1.4.0

# Cryptography
cryptography>=41.0.3

